+++ src/js/scroll.js
@@ -1,0 +1,13 @@
+const smoothLinksScroll = document.querySelectorAll('a[href^="#"]');
+
+for (let smoothLink of smoothLinksScroll) {
+  smoothLink.addEventListener('click', e => {
+    e.preventDefault();
+    const id = smoothLink.getAttribute('href');
+
+    document.querySelector(id).scrollIntoView({
+      behavior: 'smooth',
+      block: 'start',
+    });
+  });
+}
