+++ src/main.js
@@ -1,13 +1,2 @@
-const smoothLinksScroll = document.querySelectorAll('a[href^="#"]');
-
-for (let smoothLink of smoothLinksScroll) {
-  smoothLink.addEventListener('click', e => {
-    e.preventDefault();
-    const id = smoothLink.getAttribute('href');
-
-    document.querySelector(id).scrollIntoView({
-      behavior: 'smooth',
-      block: 'start',
-    });
-  });
-}
+import './js/header.js';
+import './js/scroll.js';
